+++ CHANGELOG.md
@@ -10,12 +10,11 @@ nav_order: 1
 
 <!--## [MAJOR.MINOR.PATCH] - YYYY-MM-DD -->
 
-- Upgraded `go` version to `1.24`
-- Add `aiven_organization_address` resource & datasource
-
 ## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
 
+- Upgraded `go` version to `1.24`
 - Change `aiven_kafka_topic` field `config.message_format_version` (enum): add `4.1-IV0`, `4.1`
+- Add `aiven_organization_address` resource & datasource
 
 ## [4.37.0] - 2025-03-12